feat: CSS for user forms
