+++ client/components/Login.js
@@ -3,24 +3,7 @@ import { connect } from 'react-redux';
 import { login } from '../store/auth';
 import { Container, Card, Button, Form, Row, Col } from 'react-bootstrap';
 import styled from 'styled-components';
-import { StyledFormBox, StyledWrapper } from './SignUp';
-
-// Styled Components
-export const StyledLoginButton = styled(Button)`
-  justify-content: center;
-  color: black;
-  background-color: white;
-  border: solid #d9d9d9; 1px;
-  margin-bottom: 20px;
-  margin-top: 10%;
-  margin-left: 50px;
-  &:hover {
-    color: #f5db8b;
-    border: solid #f5db8b 1px;
-  }
-`;
-
-const StyledFormLink = styled.a``;
+import { StyledFormBox, StyledWrapper, StyledLoginButton } from './styles';
 
 const Login = (props) => {
   const { name, displayName, handleSubmit, error } = props;
@@ -52,7 +35,7 @@ const Login = (props) => {
                       controlId="formBasicUserName"
                     >
                       <Form.Label column sm={2}>
-                        Email
+                        Email:
                       </Form.Label>
                       <Col sm={10}>
                         <Form.Control name="email" type="email" />
