+++ client/components/styles.js
@@ -172,6 +172,56 @@ const StyledTable = styled(Table)`
   }
 `;
 
+// Register and Signup forms
+const StyledSignupButton = styled(Button)`
+  justify-content: center;
+  color: black;
+  background-color: white;
+  border: solid #d9d9d9; 1px;
+  margin-bottom: 20px;
+  margin-top: 15%;
+  margin-left: 50px;
+
+  &:hover {
+    color: #f5db8b;
+    border: solid #f5db8b 1px;
+  }
+`;
+
+const StyledWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  align-content: center;
+  padding: 3em;
+  height: 100vh;
+  background-image: url('https://images.unsplash.com/photo-1625600243103-1dc6824c6c8a?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1064&q=80');
+  background-size: cover;
+  background-position: center;
+`;
+
+const StyledFormBox = styled.div`
+  display: flex;
+  justify-content: center;
+  border: 5px solid #b48139;
+  background-color: #ebe8eb;
+  align-text: center;
+`;
+
+const StyledLoginButton = styled(Button)`
+  justify-content: center;
+  color: black;
+  background-color: white;
+  border: solid #d9d9d9; 1px;
+  margin-bottom: 20px;
+  margin-top: 10%;
+  margin-left: 50px;
+  &:hover {
+    color: #f5db8b;
+    border: solid #f5db8b 1px;
+  }
+`;
+
 // Exports
 export {
   // Home
@@ -195,4 +245,9 @@ export {
   // AdminProducts AdminUsers
   StyledButton,
   StyledTable,
+  // Register Login
+  StyledSignupButton,
+  StyledWrapper,
+  StyledFormBox,
+  StyledLoginButton,
 };
